+++ feature-authentication/src/main/java/com/melonhead/feature_authentication/di/FeatureAuthenticationModule.kt
@@ -5,6 +5,7 @@ import com.melonhead.data_authentication.di.DataAuthenticationModule
 import com.melonhead.data_user.di.UserServiceModule
 import com.melonhead.feature_authentication.AuthRepository
 import com.melonhead.feature_authentication.AuthRepositoryImpl
+import com.melonhead.feature_authentication.navigation.LoginScreenResolver
 import com.melonhead.lib_app_events.di.LibAppEventsModule
 import com.melonhead.lib_notifications.di.LibNotificationsModule
 import kotlinx.coroutines.CoroutineScope
@@ -21,4 +22,5 @@ val FeatureAuthenticationModule = module {
     includes(UserServiceModule)
 
     single<AuthRepository> { AuthRepositoryImpl(get(), get(), get(), get(), get(), get(), get()) }
+    single<LoginScreenResolver>(createdAtStart = true) { LoginScreenResolver() }
 }
