+++ feature-authentication/src/main/res/drawable/mangadex_v2_svgrepo_com.xml
