+++ lib-navigation/src/main/java/com/melonhead/lib_navigation/keys/ActivityKey.kt
@@ -3,7 +3,7 @@ package com.melonhead.lib_navigation.keys
 import android.os.Bundle
 
 sealed class ActivityKey {
-    object MainActivity : ActivityKey()
+    data object MainActivity : ActivityKey()
     data class WebViewActivity(val params: Bundle) : ActivityKey() {
         companion object {
             const val PARAM_MANGA = "manga"
