+++ lib-navigation/src/main/java/com/melonhead/lib_navigation/keys/ScreenKey.kt
@@ -1,4 +1,5 @@
 package com.melonhead.lib_navigation.keys
 
 sealed class ScreenKey {
+    data class LoginScreen(val onLoginTapped: (email: String, password: String) -> Unit) : ScreenKey()
 }
